project1/requirements.txt
Flask
Flask-Session
psycopg2
SQLAlchemy
flask
python-dotenv
authlib
six
